Export Text from the Description Field of Feats and Traits
Request for optional description of Feats and Traits. 
With this code snippet all the Description-Field Text in Foundrys "Descrption"-Field of Feats will be displayed and "printed" to the Exported PDF

Maybe a Toggle-Option would be nice and/or to have certain Keywords or a kind of markdown for specific areas of the Descrition-Field, so that only a shorter, customized Text will be displayed insted of the whole text in the Description-Field. 
The idea of short explanations for skills and feats is inspired of the fastcharacter creation website. The sheet is very compact and there is a lot of info paked in it.

Furthermore it would be nice to make an export template for a similar design of that from fastcharacter.com.