Update to V10 mappings
With the massive changes in data structures released as part of V10, the previous version of this mapping no longer worked. This change updates the mappings to reflect those data structure changes and restore functionality.

Please note that Power Shifts will not be populated onto the sheet in the Power Shifts section, as this was getting way too finicky. They are instead added to the special abilities area, and players/GMs who wish to do so can easily translate them over.